refactor: rename package to client
